fixing events features and adding email blast

+++ my-ticket-backend/src/services/events.service.ts
@@ -8,12 +8,16 @@ export class EventService {
 
 
   async findAll(query: EventQuery) {
-    const {search, location, category, page = 1, limit = 10} = query
+    const {search, description, location, category, imageURL, page = 1, limit = 10} = query
 
     const where : any = {}
 
     if (search) {
         where.title = { contains: search, mode: 'insensitive' };
+    }
+
+    if (description) {
+      where.description = description
     }
 
     if (location) {
@@ -24,6 +28,10 @@ export class EventService {
         where.category = category
     }
 
+    if (imageURL) {
+      where.imageURL = imageURL
+    }
+
     return prisma.event.findMany({
         where,
         skip: (page - 1) * limit,
@@ -31,11 +39,27 @@ export class EventService {
     }) 
   }
 
-  async update(id: number, data: Partial<EventInput>) {
+  async verifyEventOwnership(id: number, userId: number): Promise<boolean> {
+    const event = await prisma.event.findUnique({
+        where: { id },
+        select: { userId: true }
+    });
+    
+    return event?.userId === userId;
+}
+
+  async update(id: number, userId: number, data: Partial<EventInput>) {
+    const isOwner = await this.verifyEventOwnership(id, userId);
+        if (!isOwner) {
+            throw new Error("You are not authorized to update this event");
+        }
     return prisma.event.update({ where: { id }, data: {...data, updatedAt: new Date()} })
   }
 
-  async delete(id: number) {
+  async delete(id: number, userId: number) {
+
+    await this.verifyEventOwnership(id, userId)
+
     return prisma.event.delete({ where: { id } })
   }
 }
